backend/requirements/base.txt
# Core Django
Django>=5.1,<5.2
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-gis>=1.1,<2.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.0,<3.0

# GeoDjango - match installed libgdal version (3.10.x in debian slim)
GDAL>=3.8,<3.11

# Celery for async tasks
celery>=5.3,<6.0
redis>=5.0,<6.0

# Web scraping
selenium>=4.15,<5.0
beautifulsoup4>=4.12,<5.0
lxml>=5.0,<6.0

# Translation
deepl>=1.17,<2.0

# Geolocation
geoip2>=4.7,<5.0

# Image handling
Pillow>=10.1,<11.0

# Calendar export
icalendar>=5.0,<6.0

# Rich text editor
django-prose-editor[sanitize]>=0.22,<1.0

# Utilities
python-dotenv>=1.0,<2.0
python-dateutil>=2.8,<3.0